--- a/pom.xml
+++ b/pom.xml
@@ -209,6 +209,26 @@
 			</plugin>
 
 		</plugins>
+
+		<pluginManagement>
+			<plugins>
+				<plugin>
+					<groupId>com.gradle</groupId>
+					<artifactId>gradle-enterprise-maven-extension</artifactId>
+					<configuration>
+						<gradleEnterprise>
+							<normalization>
+								<runtimeClassPath>
+									<ignoredFiles>
+										<ignoredFile>builddef.lst</ignoredFile>
+									</ignoredFiles>
+								</runtimeClassPath>
+							</normalization>
+						</gradleEnterprise>
+					</configuration>
+				</plugin>
+			</plugins>
+		</pluginManagement>
 	</build>
 
 	<repositories>