--- a/spring-data-jpa/src/main/java/org/springframework/data/jpa/provider/HibernateUtils.java
+++ b/spring-data-jpa/src/main/java/org/springframework/data/jpa/provider/HibernateUtils.java
@@ -26,6 +26,8 @@ import org.springframework.lang.Nullable;
  * @author Oliver Gierke
  * @author Mark Paluch
  * @author Jens Schauder
+ * @author Donghun Shin
+ * @author Greg Turnquist
  * @since 1.10.2
  * @soundtrack Benny Greb - Soulfood (Live, https://www.youtube.com/watch?v=9_ErMa_CtSw)
  */
@@ -54,8 +56,8 @@ public abstract class HibernateUtils {
 
 		// Try the old way, as it still works in some cases (haven't investigated in which exactly)
 
-		if (query instanceof Query) {
-			return ((Query<?>) query).getQueryString();
+		if (query instanceof Query<?> hibernateQuery) {
+			return hibernateQuery.getQueryString();
 		} else {
 			throw new IllegalArgumentException("Don't know how to extract the query string from " + query);
 		}