--- a/spring-data-jpa/src/main/java/org/springframework/data/jpa/repository/query/Meta.java
+++ b/spring-data-jpa/src/main/java/org/springframework/data/jpa/repository/query/Meta.java
@@ -27,6 +27,7 @@ import org.springframework.util.StringUtils;
  * Value object to hold metadata about repository methods.
  *
  * @author Greg Turnquist
+ * @author Donghun Shin
  * @since 3.0
  * @see org.springframework.data.jpa.repository.Meta
  */