--- a/spring-data-jpa/src/main/java/org/springframework/data/jpa/repository/query/QueryUtils.java
+++ b/spring-data-jpa/src/main/java/org/springframework/data/jpa/repository/query/QueryUtils.java
@@ -80,6 +80,7 @@ import org.springframework.util.StringUtils;
  * @author Simon Paradies
  * @author Vladislav Yukharin
  * @author Chris Fraser
+ * @author Donghun Shin
  */
 public abstract class QueryUtils {