```
Summary
-------
Project:                         Spring Data JPA Parent
Git repo:                        file:///Users/eric/Projects/spring-projects/spring-data-jpa
Git branch:                      erichaagdev/main
Git commit id:                   770a435470ede2b306f9eca8070114d2f5d383a7
Git options:                     --depth=1
Project dir:                     <root directory>
Maven goals:                     dependency:list test
Maven arguments:                 -Dsort -U -B
Experiment:                      02 Validate local build caching - different project locations
Experiment id:                   exp2-maven
Experiment run id:               64ef647a
Experiment artifact dir:         .data/02-validate-local-build-caching-different-locations/20230830T104706-64ef647a
Build scan first build:          https://ge.solutions-team.gradle.com/s/b2asglgpzmfrq
Build scan second build:         https://ge.solutions-team.gradle.com/s/qu2awwqd7dsas

Performance Characteristics
---------------------------
Initial build time:              2m 47.705s
Build time with instant savings:     5.798s, 2m 41.907s savings
Build time with pending savings:     5.798s, 0.000s additional savings
Avoided cacheable goals:         22 goals, 2m 37.942s total saved execution time
Executed cacheable goals:         0 goals, 0.000s total execution time
Executed non-cacheable goals:    28 goals, 2.806s total execution time
Serialization factor:            1.0x first build, 1.0x second build

See https://gradle.com/bvs/main/Maven.md#performance-characteristics for details.

Investigation Quick Links
-------------------------
Goal execution overview:         https://ge.solutions-team.gradle.com/s/qu2awwqd7dsas/performance/execution
Executed goals timeline:         https://ge.solutions-team.gradle.com/s/qu2awwqd7dsas/timeline?outcome=success,failed&sort=longest
Avoided cacheable goals:         https://ge.solutions-team.gradle.com/s/qu2awwqd7dsas/timeline?outcome=from-cache&sort=longest
Executed cacheable goals:        https://ge.solutions-team.gradle.com/s/qu2awwqd7dsas/timeline?cacheability=cacheable&outcome=success,failed&sort=longest
Executed non-cacheable goals:    https://ge.solutions-team.gradle.com/s/qu2awwqd7dsas/timeline?cacheability=any-non-cacheable&outcome=success,failed&sort=longest
Build caching statistics:        https://ge.solutions-team.gradle.com/s/qu2awwqd7dsas/performance/build-cache
Goal inputs comparison:          https://ge.solutions-team.gradle.com/c/b2asglgpzmfrq/qu2awwqd7dsas/goal-inputs?cacheability=cacheable
```